Fixes COP attribute form.
The COP attribute form is rendered to a string, that string is then
escaped as JavaScript which is then placed on the page in a script tag
and is set to a variable. The problem was that this string was getting
a newline inserted in it for some reason. It appears that Haml was
splitting this long line of JavaScript for some reason, we think
because it contained a “>” character. When we switched the form to ERB
the issue disappeared. This will likely affect the COP file form as
well.