Fixes missing blank options in form selects.
The default form builder at app/builders/default_form_builder.rb
destructively removes the last item from the incoming arguments array
if it’s a hash and passes it on to the label helper. The problem is
that since this is destructive, when it calls super to the original
helper method and passes the arguments along they no longer contain the
options hash and so the underlying helper seemingly ignores any passed
options. Since this behaviour might actually be relied on in other
parts of the code I am choosing the remedy the issue in a case-by-case
basis for fear of introducing regressions.