reload contacts relationships to avoid Active Record creating duplicated join objects
This happens because we use Session to marshal and restore the object between each step and it appears that since rails 4 this causes problems in maintaining proper many to many relationships. we should move away from this behaviour and persist the intermediate objects in the database.
Hopefully it will happen as a part of the public site rebuild

see https://www.pivotaltracker.com/n/projects/26398/stories/87126256