Rename PageForm to Layout.
I feel like this really helps clear things up. Containers no longer
have a kind, they have a layout. Layouts are what define the shape of
the JSON payload a container has, and they also help determine what
template to use for rendering the page for consumers.

We now have the following structure:

Container (Model)
  Layout (Domain Class)
    Containers belong to one layout through the layout enum propery
    on Container. Layout is used at design time for verifying the
    shape of the JSON data stored in Payload records associated with
    the Container.
  Payload (Model)
    Containers have n payloads, they store the JSON data that is
    validated by the container's associated Layout class.
  Page (Domain Class)
    Page is used as a presenter to connect everything together for
    use in the consumer interface.

It is possible that Layout + Page might be able to live together in
the future.