Containers can now be published.
I think it might be time for a service object here,
ContainerPublishingService or something, we’ll see how this feels in
the future.

How does it work?

You always edit the “draft_payload”, when you click publish it copies
the draft payload (inserts a new row in the payloads table). I then
takes the id of that new row and sets it to be the “public_payload” of
the container. If the container already has a public payload, it’s JSON
data is compared to the draft’s JSON data and only executes the
publishing workflow if the data payloads differ.