Completed sitemap tree editing semantics.
When a container:

Is destroyed then its children are re-associated to the destroyed
containers parent container (or root if there is no parent).

Is moved (its parent_container_id changes) then it’s old parent is
notified of the change to recache it’s immediate child count.

Is added as a child its parent container is notified of the change to
recalculate its immediate child container counter cache.