[WIP] Add event and tagging presenter test
This commit includes tests for the edge case where sponsors and tagging
were not being returned as selected on validation fail. The test for
sponsors and topics pass as they should but the tests for issues and
sectors fail. I can’t seem to figure out why. Mind having a look?