Handle missing default templates as 404s
https://app.honeybadger.io/projects/43892/faults/13657502

Paths like /News which will match a container by path and yet were not meant to
be shown via the catch_all route will cause a server error when the template is
not found.

This change adds a check to see if the default template for a layout exists,
rendering a 404 if it doesn't.