Replace silently broken `assert_select, href:` syntax
This style of assertion was mentioned in a few articles as an nice way
to write such a test but it actually does not work and does not thrown
an error. It’s being replaced with the style found in
ActionDispatch::Assertions::SelectorAssertions documentation. Also
fixes an assertion in assert_render_resources_content_block_component
in the integration test helper that apparently wasn’t written properly.
Closes #317.