Expulsion dates
Organization#cop_due_on holds the value for the next cop or coe regardless of business type.  It is calculated before being saved in the database with the correct offset for the type (1 yr for COP, 2 yrs for COE).

This commit replaces coe_due_on with communication_due_on which is simply an alias for cop_due_on. At least we carry the intent this way.

We also introduce Organization#projected_expulsion_date which is 1 year past the communication_due_on date.

Finally we DRY up the expulsion period into EXPULSION_THRESHOLD.