Not Applicable listing status is no longer available to contact points
Business participants are required to give their listing/ownership type (public, private, subsidiary or state-owned) during the application process, but we noticed that "Not Applicable" is given as an option when participants login to their profile and choose to update their info. Can we make it so that "Not Applicable" is not an available option for participant contact point users to select?