Remove home page caching.
Home page caching was interfearing with news and event publishing.
Because the cache keys were based on container and payload, new events
and news wouldn't force the cache to rebuild.

It is not clear from the callsite (the controller) that data other than
container and payload is required, the data is loaded further down the
stack in the components.

A new caching strategy could be created once all of the dependencies are
clear.

For now, we'll disable the cache and monitor the performance hit on home
page times.