Remove delisting_on function from COP mailer
This old code path was causing upcoming delisting warnings to report
dates in the past. The SME Moratorium was the root cause and as it is no
longer relevant, we're removing the code.

DC RMS Ltd (2997)
inactive_on: 30 Mar 2007
delisting_on: 7 OCT 2016
SME

reported in it's 30 day warning email that it's next COP was due on 30
Mar 2008.