Remove directory indexes from shared/public
* only allow webdav to public/pics and public/docs
* turn off indexes and robots to the shared directory

Google has been indexing webdav content and surfacing private documents.
For example attachments for LogoRequest comments such as:
/admin/organizations/5964/logo_requests/3172

These documents are not intended to be public but as they live in the
public folder... they are. This change only mitigates the issue by
preventing them from being indexed.

We leave webdav open for /docs and /pics as they are intended to be
publicly sharable.

Attachment such as these which are only intended for admin use should be
moved out of the public document root to avoid accidental exposure.