No longer use RedisSessionStore (#431)
Replace the redis session store with the default json/cookie store.
The impetus here is to allow rack-request_replication to work so that we
can replicate production traffic to staging in an effort to better test
against live traffic.

rack-request_replication doesn't work with redis backed session so we're
migrating away.

Participant signups rely on a large session space to store the pending
signup data across requests. This is too big for the cookie store to
handle and will overflow the 4k alotment.

PendingSignup is introduced here which continues to store these signups
in redis.