Fix COP Files
COPs in draft state were not able to be published as any empty COP files
created when saving the draft would be invalid when the publish was
attempted. Further, the 'Remove' checkbox was hidden and not usable to
fix this situation. Finally, the form would add an additional empty cop
file even if the form already has one.

* Only show the default file field if they aren't an existing files
* Fix styling on individual file entry fields
* Don't persist CopFiles that don't have an attachment