RemovalReason.for_filter no longer alters it's arguments
Using .map! on the method arguments will mutate the values given, causing undesired effects outside the scope of the method.

For example, if the arguments are an unfrozen constant, those values are now globally changed.

use the regular .map to make a copy and avoid the aliasing.

Also freezes a bunch of constants :]