Redirect users who don't need to update their passwords
Some users are being asked to update their password immediately after a
forgotten password change. The pick a new password form requires them to choose
a strong, valid password, but for some reason last_password_change_at is not
being given a value and they are being asked again to change their password.

Having seen evidence of the problem, but not being able to reproduce it, I'm
adding this redirect to move users to their dashboard page if they are
mistakenly asked to change their password yet again.