Re-write login with cookie tests
Changes to more recent versions of Devise will break with the current test method.

Refactor to use time travelling to describe the sequqnce of events and let devise sort the reset out.

It should be noted that Devise expects the time that the cookie was generated to be after the time .remember_me! was called. This is now included in the test.