Wrap Campaign and Contribution sync ops in a transaction
We have been seeing errors about duplicate entries on insert of
contributions. I believe this is a result of racey behaviour but I
haven't been able to prove that for certain. Wrapping the read and write
in a transaction should fix that if that is indeed the problem and
should probably have been there from the start.