Don't make network calls in Salesforce tests
Previously, tests other than the UngcSyncTests were making real http
calls in test mode.  Returning a canned "ok" response for all network
calls in test mode ment skipping the error paths and dropped the test
coverage below SF's minimum 75%.

Adding a static mocked flag is not a great solution, but gets the job
done so that EchoTS can deploy at least.

In the near future, I hope to replace this whole Apex solution with two
Salesforce PushTopics for Contribution and Campaigns and a Ruby listener
that will do most of the work. We can thereby skip all of this
complexity in SF and bring these concerns back to the Ruby side
(mostly).