Joiner form fixes (#623)
* Remove unused foundation_contact option

* Stop the signup when financial contact is invalid

No longer silently fails when financial contact is invalid. This used to
be okay somehow in the past but we now require a financial contact.

Turn transaction failures into messages the user can see/correct.

* Ensure there is a session started for signups.

Previously, if there wasn't anything already written to the session,
there would be no session id and therefore every page of the signup
would start fresh. This isn't a problem until the user hits a
validation error and sends the whole thing back to step1.

By ensuring there is something in the session, it can be reused as it
was meant to be.

* Check for duplicate emails among unsaved contacts

* Don't try to save any action platforms if none are selected