Simple Salesforce listener (#818)
Implements a simple Salesforce PushTopic listener which runs as a daemon that is monitored by God. This listener does nothing other than to log events to the log file. It is intended to serve as the basis for future work.

The listener will re-connect itself to the topic every 30 minutes to work around Salesforce disconnecting clients every 2 hours and faye not being able to reconnect.

* Use fork of cookiejar to skip cookie domain checks
Salesforce sends the wrong cookie domain. We use this fork to skip the cookie domain check. This is hack-y but should be tollerable for the short term.
See https://github.com/restforce/restforce/issues/120 for more discussion.