Fix logic which defaults to non-businesses. (#848)
https://github.com/unglobalcompact/UNGC/issues/845

MicroEnterprises should be created with the business COP due period of 1
year but was instead being set to 2 years.

This was because when we introduced the COP vs COE distinction, we opted
to test against #company? which only includes SME and Companies.
Switching to business? still leaves out Micro Enterprises as they are
neither business nor non-business (They have a type_property: 0 which
indicates neither along with Signatory types). It's possible that this
isn't correct, but it's been that way since 2010, so I'm opting to
believe that is intentional.

We now check against #non_business? and default to COP which was the
behaviour before the introduction of COEs.