Add accepts_eula to signup form (#865)
https://github.com/unglobalcompact/UNGC/issues/864

We need to add the following question to the new joiner form:
"I have read and agree with the terms of the privacy policy and cookies policy".
Have the links open into a new window so they don't lose their work.
There should be a checkbox adjacent to it that must be clicked in order to submit the form.
You can place this under the Letter of Commitment upload feature on the last page of the form. Above the "Submit" button.