Fix fetch on AR.
This looks like it never worked properly. Here we're calling .fetch on
an ActiveRecord relation, which does happen to exist but is certainly
not what I had intended. Then intent is to show any existing AP
Subscriptions in the form. As the select never returned anything
meaningful, this code would always fall through to the default block
which provided resonable behaviour.

Trading #fetch in for a meaningful call, #where in this case so we can
get the subscription matching the platform, exposes the 2nd problem
which is that the caller is expecting instances of SubscriptionForm.

So now we look for the first subscription, wrapping it if it's present
or providing a empty default explicitly.

The implementation could certainly be more efficient, but getting it to
work properly is good enough for now (without #fetch which disappears in Rails 5.1).